Sparse: Implement simple unary ufuncs operators
By "simple" I mean operators that map 0->0 so we can implement it by
just re-dispatching on the values tensor. That does mean we have `sin`
but not `cos` for example, but without fill value support this is the
best that can be done.

Most of these don't support autograd because the derivative formulas
use unsupported operators.

ghstack-source-id: 10b37350d6dbe2e14c2b6a02a8bcbe0b6e86d54c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/68887